Update task locations to make more sense
